requirements.txt
# Core dependencies for AI Prompt Generator
streamlit==1.28.1
requests>=2.31.0

# Additional utilities
python-dotenv==1.0.0
pydantic==2.5.0
typing-extensions==4.8.0

# UI enhancements
streamlit-option-menu==0.3.6
streamlit-extras==0.3.6

# Data handling - using compatible versions for Windows
pandas>=2.0.0
numpy>=1.24.0

# File operations
pyyaml==6.0.1
